Refactor hero retrieval logic by creating a separate heroes data module and service
