+++ 03-TypeScript/src/data/heroes.ts
@@ -1,0 +1,23 @@
+interface Hero {
+    id: number;
+    name: string;
+    owner: string;
+}
+
+export const heroes: Hero[] = [
+    {
+        id : 1,
+        name: 'Ironman',
+        owner: 'Marvel',
+    },
+    {
+        id : 2,
+        name: 'Spiderman',
+        owner: 'Marvel',
+    },
+    {
+        id : 3,
+        name: 'Batman',
+        owner: 'DC',
+    },
+];
