+++ 03-TypeScript/src/services/hero.service.ts
@@ -1,0 +1,5 @@
+import { heroes } from "../data/heroes";
+
+const findHeroById = ( id: number ) => {
+    return heroes.find( ( hero ) => hero.id === id);
+};
